# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deploy containers to cloud servers. Push a config YAML + env, get an ordered deployment plan, execute it.",
        "title": "nvoi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workspaces/{workspace_id}/repos/{repo_id}/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the latest config version for a repo. Env is hidden unless ?reveal=true.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get latest config",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Repo ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Show env values",
                        "name": "reveal",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.configResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pushes a new versioned config snapshot. Validates YAML, expands managed services, validates env references, and persists credentials.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Push config",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Repo ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Config payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.pushConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.RepoConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/repos/{repo_id}/config/plan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the ordered deployment plan for the latest config, including both delete and set steps.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get execution plan",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Repo ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.planResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/repos/{repo_id}/configs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all config versions for a repo, newest first. Env is stripped.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "List config versions",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Repo ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api_handlers.configListItem"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_getnvoi_nvoi_internal_api.BuildProvider": {
            "type": "string",
            "enum": [
                "local",
                "daytona",
                "github"
            ],
            "x-enum-varnames": [
                "BuildLocal",
                "BuildDaytona",
                "BuildGitHub"
            ]
        },
        "github_com_getnvoi_nvoi_internal_api.ComputeProvider": {
            "type": "string",
            "enum": [
                "hetzner",
                "aws",
                "scaleway"
            ],
            "x-enum-varnames": [
                "ComputeHetzner",
                "ComputeAWS",
                "ComputeScaleway"
            ]
        },
        "github_com_getnvoi_nvoi_internal_api.DNSProvider": {
            "type": "string",
            "enum": [
                "cloudflare",
                "aws"
            ],
            "x-enum-varnames": [
                "DNSCloudflare",
                "DNSAWS"
            ]
        },
...
        "github_com_getnvoi_nvoi_internal_api.RepoConfig": {
            "type": "object",
            "properties": {
                "build_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.BuildProvider"
                },
                "compute_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.ComputeProvider"
                },
                "config": {
                    "description": "YAML",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dns_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.DNSProvider"
                },
                "id": {
                    "type": "string"
                },
                "repo_id": {
                    "type": "string"
                },
                "storage_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.StorageProvider"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_getnvoi_nvoi_internal_api.StorageProvider": {
            "type": "string",
            "enum": [
                "cloudflare",
                "aws"
            ],
            "x-enum-varnames": [
                "StorageCloudflare",
                "StorageAWS"
            ]
        },
...
        "internal_api_handlers.configListItem": {
            "type": "object",
            "properties": {
                "build_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.BuildProvider"
                },
                "compute_provider": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.ComputeProvider"
                        }
                    ],
                    "example": "hetzner"
                },
                "config": {
                    "type": "string"
                },
                "dns_provider": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.DNSProvider"
                        }
                    ],
                    "example": "cloudflare"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "storage_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.StorageProvider"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_api_handlers.configResponse": {
            "type": "object",
            "properties": {
                "build_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.BuildProvider"
                },
                "compute_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.ComputeProvider"
                },
                "config": {
                    "description": "YAML",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dns_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.DNSProvider"
                },
                "env": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "repo_id": {
                    "type": "string"
                },
                "storage_provider": {
                    "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.StorageProvider"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "workspace not found"
                }
            }
        },
...
        "internal_api_handlers.planResponse": {
            "type": "object",
            "properties": {
                "steps": {
                    "description": "Steps is the ordered deployment sequence.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_handlers.planStep"
                    }
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_api_handlers.planStep": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string",
                    "example": "instance.set"
                },
                "name": {
                    "type": "string",
                    "example": "master"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "internal_api_handlers.pushConfigRequest": {
            "type": "object",
            "required": [
                "compute_provider",
                "config"
            ],
            "properties": {
                "build_provider": {
                    "enum": [
                        "local",
                        "daytona",
                        "github"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.BuildProvider"
                        }
                    ],
                    "example": "github"
                },
                "compute_provider": {
                    "enum": [
                        "hetzner",
                        "aws",
                        "scaleway"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.ComputeProvider"
                        }
                    ],
                    "example": "hetzner"
                },
                "config": {
                    "type": "string",
                    "example": "servers:\n  master:\n    type: cx23\n    region: fsn1"
                },
                "dns_provider": {
                    "enum": [
                        "cloudflare",
                        "aws"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.DNSProvider"
                        }
                    ],
                    "example": "cloudflare"
                },
                "env": {
                    "type": "string",
                    "example": "HETZNER_TOKEN=xxx\nRAILS_MASTER_KEY=yyy"
                },
                "storage_provider": {
                    "enum": [
                        "cloudflare",
                        "aws"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_internal_api.StorageProvider"
                        }
                    ],
                    "example": "cloudflare"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token from POST /login. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}